Split the coefficient functions in a header and a source file.
Change the coefficient function signatures such that they return a pair of integer (rational number) such that it can be compiled to a library with a simple API.

Add a library compilation target for the coefficients.

Add Utility functions for handling conversion between numerator and denominator and Expressions.

Place coefficients and everything below it, i.e. the library, in a namespace.